+++ src/Components/Attendees.js
@@ -1,3 +1,18 @@
-export default function Attendees() {
-  return;
+
+import Attendee from "./Attendee";
+
+export default function Attendees({attendees, event, updateEventAttendance}) {
+  
+  return(
+    <div className="attendees">
+      {attendees.map((attendee, index) => (
+          <Attendee
+            event={event}
+            attendee={attendee}
+            updateEventAttendance={updateEventAttendance}
+          />
+      ))}
+    </div>
+  )
+  ;
 }
