+++ src/Components/Footer.js
@@ -1,3 +1,13 @@
 export default function Footer() {
-  return;
+  return(
+    <>
+    <footer>
+      <ul>
+        <li>Contact</li>
+        <li>About</li>
+        <li>Legal</li>
+      </ul>
+    </footer>
+  </>
+  );
 }
